+++ src/providers/webModules.js
@@ -1,25 +1,36 @@
-import { readdirSync } from 'fs';
+import { readdirSync, lstatSync } from 'fs';
 import { join, extname } from 'path';
 import slash from 'slash';
 
 import { createCompletionList, resolve } from '../util';
 
-function getFiles (path, ext) {
+const isJsw = (path) => extname(path) === '.jsw';
+const isDir = (path) => lstatSync(path).isDirectory();
+
+const getItems = (path) => {
   const fileList = [];
   const files = readdirSync(path);
 
   files.forEach((file) => {
     const filePath = join(path, file);
+    const name = filePath.slice(path.length + 1);
 
-    if (extname(file) === ext) {
-      const name = filePath.slice(path.length + 1);
-
-      fileList.push(slash(name));
+    if (isJsw(file)) {
+      fileList.push({
+        name: slash(name),
+        kind: 16,
+        detail: 'Corvid Web Module',
+      });
+    } else if (isDir(filePath)) {
+      fileList.push({
+        name: slash(name),
+        kind: 18,
+      });
     }
   });
 
   return fileList;
-}
+};
 
 export const jsw = {
   provideCompletionItems(doc, position) {
@@ -27,17 +38,13 @@ export const jsw = {
     const match = /^(?:import.+)(?:['"])(backend\/.*)/m.exec(prefix);
 
     if (Array.isArray(match)) {
-      const path = resolve('src/backend');
-      const files = getFiles(path, '.jsw');
+      try {
+        const dir = match[1].split('/').slice(0, -1).join('/');
+        const path = resolve('src', dir);
+        const items = getItems(path);
 
-      const items = files.map((name) => {
-        return {
-          name,
-          kind: 16,
-        };
-      });
-
-      return createCompletionList(items);
+        return createCompletionList(items);
+      } catch (error) { /**/ }
     }
   },
 };